feat: add JWT authentication

+++ src/main/java/com/example/demo/controller/WaveController.java
@@ -3,8 +3,12 @@ package com.example.demo.controller;
 import com.example.demo.dto.wave.WaveCreateDTO;
 import com.example.demo.dto.wave.WaveResponseDTO;
 import com.example.demo.entity.Ponto;
+import com.example.demo.entity.User;
 import com.example.demo.entity.Wave;
+import com.example.demo.exception.exceptions.UnauthorizedException;
+import com.example.demo.jwt.JwtUserDetails;
 import com.example.demo.service.PontoService;
+import com.example.demo.service.UserService;
 import com.example.demo.service.WaveService;
 import com.example.demo.utils.Calculadora;
 import jakarta.transaction.Transactional;
@@ -12,6 +16,7 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -28,11 +33,15 @@ public class WaveController {
     private final WaveService waveService;
     private final PontoService pontoService;
     private final Calculadora calculadora;
+    private final UserService userService;
 
     @PostMapping
     @Transactional
-    public ResponseEntity<Long> create(@RequestBody @Valid WaveCreateDTO dto) {
+    public ResponseEntity<Long> create(@RequestBody @Valid WaveCreateDTO dto,
+                                       @AuthenticationPrincipal JwtUserDetails userDetails) {
+        User user = userService.findById(userDetails.getId());
         Wave wave = new Wave(dto);
+        wave.setUser(user);
 
         // Salva a simulação e obtém o ID gerado
         Long simulacaoId = waveService.save(wave);
@@ -47,6 +56,7 @@ public class WaveController {
             pontoService.save(ponto);
         }
 
+        // OBS: diferente do convencional, esta rota retorna apenas o ID da entidade criada para facilitar no excel
         return ResponseEntity.status(HttpStatus.CREATED).body(simulacaoId);
     }
 
@@ -60,12 +70,20 @@ public class WaveController {
     @GetMapping
     public ResponseEntity<List<WaveResponseDTO>> getWaves() {
         List<Wave> ondas = waveService.findAll();
-        List<WaveResponseDTO> responseDtos = ondas.stream().map(WaveResponseDTO::new).toList();
-        return ResponseEntity.ok(responseDtos);
+        List<WaveResponseDTO> responses = ondas.stream().map(WaveResponseDTO::new).toList();
+        return ResponseEntity.ok(responses);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteWave(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteWave(@PathVariable Long id,
+                                           @AuthenticationPrincipal JwtUserDetails userDetails) {
+        User user = userService.findById(userDetails.getId());
+        Wave wave = waveService.findById(id);
+
+        if (!wave.getUser().equals(user) && !User.Role.ROLE_ADMIN.equals(user.getRole())) {
+            throw new UnauthorizedException("Você não tem permissão para deletar esta simulação");
+        }
+
         waveService.deleteWave(id);
         return ResponseEntity.noContent().build();
     }
